Add directive for advancing to the next question on selecting an answer
* Instead of cluttering the recordAnswer directive with logic for
advancing to the next question, this has been split out into it's own
chunk of code. Because it depends on DOM interaction, a directive feels
like the right choice.
* Changing questions now changes the active state rather than clicking
on the question container. This will provide the flow through all the
questions and ensure everything is filled in correctly.
* With the changeQuestion function, it's now trivial to specify a DOM
element and a number to update the current question to. This will allow
advancing through questions as well as going backwards.
* Removed some unnecessary questions from the model.